+++ examples/async_std.rs
@@ -70,7 +70,7 @@
 //         });
 //     }));
 //
-//     while app.wait() {
+//     while app.wait().unwrap() {
 //         match r.try_next() {
 //             Ok(Some(Message::Inc(v))) => frame.set_label(&format!("increments by {}", v)),
 //             _ => (),