+++ src/lib.rs
@@ -127,7 +127,7 @@
 //!     but_inc.set_callback(Box::new(move || s.send(Message::Increment)));
 //!     but_dec.set_callback(Box::new(move || s.send(Message::Decrement)));
 //!
-//!     while app.wait() {
+//!     while app.wait().unwrap() {
 //!         let label: i32 = frame.label().parse().unwrap();
 //!         match r.recv() {
 //!             Some(Message::Increment) => frame.set_label(&(label + 1).to_string()),